task#35416a

+++ app/src/main/java/com/example/soccert/ui/home/HomeViewModel.kt
@@ -45,7 +45,7 @@ class HomeViewModel(
 
     fun getEventByDateAndLeague(fromDate: String, toDate: String) {
         _itemCompetition.value?.let { item ->
-            soccerRepository.getEvents(item.leagueID.toInt(), fromDate, toDate)
+            soccerRepository.getEvents(item.leagueID, fromDate, toDate)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe({
